feat(swagger): migrated schemas to schemas.js
